Use native xlib window queries when glXQueryDrawable fails
This fixes #9
GLX 1.4 is not clear wether glXQueryDrawable should work for bare Window
XID-s. Seems it works on nVidia blob, but fails on all Mesa drivers.

The fix here is to intercept GLX error, and continue only with native
X11 window queries. This may now work for GLXPixmaps/GLXPBuffers, but
fixes glxgears, so seems good.