ES3 fix for depth texture view: use color internalformat for drawing the texture.
For D & D+S textures auxctx getter was broken, as it tried binding D or D+S RT to color attachment. As ES2/3 does not support reading depth or stencil from FBO, the only sane way of fixing this is to always render D or D+S data into color.

The downsite is total lack of precision. (TODO: maybe float rendering /if possible/ should be done in auxctx?)